Define SMALLEST dataset
The idea is to map datasets that take less time to enable runtime check for PR changes.
This command
```
java -jar ./optaplanner-perf-benchmark/target/optaplanner-8-benchmarks.jar -jvmArgs '-Xms6144m -Xmx6144m' -v NORMAL -foe false -gc true -wi 1 -i 1 -f 1 -p constructionHeuristicType=FIRST_FIT -p dataSet=SMALLEST 'org.jboss.qa.brms.performance.(?!.*scalability).*'
```
will run non-scalability tests in quick mode

When performance tests starts at Friday and takes main branch of OptaPlanner it can fail not only on the build but in the runtime (for example if we change paths in optaplanner-examples) to avoid that and be able to include this as a part of CI I propose to create another tag for short datasets. That allows to run all tests in runtime for short datasets. That takes 4 minutes currently on my local.

Currently, will help as to catch the issues earlier when the new version of OptaPlanner gets changed